test: 合并 notifier 四文件到 test_notifier.py
- test_notifier.py 并入 4 文件独有 31 例：extended(3) http(3+2 fixture) full(17) watching_summary(11)
- 各文件重复用例（冷却、模板替换、邮件/主题/文本、配置获取等）删除
- watching_summary 需 helper（_make_notifier/_make_watching_data）一并带入
- test_notifier_pending_candidate / test_inbox_notifications 全独有，保留
